+++ examples/manage-secrets/outputs.tf
@@ -14,7 +14,22 @@
  * limitations under the License.
  */
 
-output "contents" {
-  description = "The actual value of the requested secret"
-  value       = module.secret.contents
+output "app_buckets" {
+  value       = module.secret_storage.app-buckets
+  description = "The lists of created application-specific buckets"
+}
+
+output "shared_buckets" {
+  value       = module.secret_storage.shared-buckets
+  description = "The lists of created shared buckets"
+}
+
+output "app_secret" {
+  description = "The actual value of the requested app-secret"
+  value       = module.fetch_app_secret.contents
+}
+
+output "shared_secret" {
+  description = "The actual value of the requested shared secret"
+  value       = module.fetch_shared_secret.contents
 }