+++ modules/gcs-object/data.tf
@@ -15,12 +15,12 @@
  */
 
 data "google_storage_object_signed_url" "file_url" {
-  bucket   = "${var.bucket}"
-  path     = "${var.path}"
-  duration = "${var.duration}"
+  bucket   = var.bucket
+  path     = var.path
+  duration = var.duration
 }
 
 data "http" "remote_contents" {
-  url        = "${data.google_storage_object_signed_url.file_url.signed_url}"
+  url        = data.google_storage_object_signed_url.file_url.signed_url
   depends_on = ["null_resource.force-wait"]
 }