+++ test/fixtures/create-multiply-secrtes/outputs.tf
@@ -14,7 +14,16 @@
  * limitations under the License.
  */
 
-output "contents" {
-  description = "The actual value of the requested secret"
-  value       = module.secret.contents
+output "app_buckets" {
+  value       = module.example.app_buckets
+  description = "The lists of created application-specific buckets"
+}
+
+output "shared_buckets" {
+  value       = module.example.shared_buckets
+  description = "The lists of created shared buckets"
+}
+
+output "project_id" {
+  value = var.project_id
 }